add video length for Bilibili preview share
- also fixed bilibili.XWebView[cid] may decode failed

+++ src/main/scala/cc/sukazyo/cono/morny/extra/bilibili/XWebView.scala
@@ -26,7 +26,7 @@ case class XWebView (
 	owner: XWebView.User,
 	stat: Ignore,
 	dynamic: String,
-	cid: Int,
+	cid: Long,
 	dimension: Ignore,
 	premiere: Ignore,
 	teenage_mode: Int,
